+++ equ-api/src/main/java/equ/api/cashregister/CashRegisterHandler.java
@@ -28,7 +28,7 @@ public abstract class CashRegisterHandler<S extends SalesBatch> extends Machiner
     public abstract SalesBatch readSalesInfo(String directory, List<CashRegisterInfo> cashRegisterInfoList) throws IOException, ParseException, ClassNotFoundException;
 
     public abstract void requestSalesInfo(List<RequestExchange> requestExchangeList, Set<String> directorySet,
-                                          Set<Long> succeededRequests, Map<Long, Throwable> failedRequests, Map<Long, Throwable> ignoredRequests) throws IOException, ParseException;
+                                            Set<Integer> succeededRequests, Map<Integer, Throwable> failedRequests, Map<Integer, Throwable> ignoredRequests) throws IOException, ParseException;
 
     public abstract void finishReadingSalesInfo(S salesBatch);
